 requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
gunicorn>=20.1.0
whitenoise>=6.5.0
python-decouple>=3.8
django-cors-headers>=4.0.0
